Removed the test utils to appropriate folder
