+++ packages/smartcontracts/src/tests/BridgeProxy.supportedToken.spec.ts
@@ -3,7 +3,8 @@ import { expect } from 'chai';
 import { ethers } from 'hardhat';
 
 import { BridgeV1, TestToken } from '../generated';
-import { deployContracts, toWei } from './testHelper';
+import { deployContracts } from './testUtils/deployment';
+import { toWei } from './testUtils/mathUtils';
 
 describe('Add and Removed Supported ETH and ERC20 tokens', () => {
   async function supportedToken(proxyBridge: BridgeV1, testToken: TestToken) {
