+++ packages/smartcontracts/src/tests/testUtils/mathUtils.ts
@@ -1,0 +1,12 @@
+import { BigNumber } from 'ethers';
+import { ethers } from 'hardhat';
+
+export function toWei(amount: string): BigNumber {
+  return ethers.utils.parseEther(amount);
+}
+
+export function calculateFee(amount: BigNumber, transactionFee: BigNumber): BigNumber {
+  const feeAmount = amount.mul(transactionFee).div(10000);
+  const netAmountAfterFee = amount.sub(feeAmount);
+  return netAmountAfterFee;
+}
